refactor: consolidate UserActionTrackingService into Application layer
- Moved `UserActionTrackingService` and its interface from Web layer to Application layer.
- Updated service references in dependent projects.
- Adjusted folder structure and cleaned up unused code in the Web layer.
