+++ src/Dfe.PlanTech.Application/Services/Interfaces/IUserActionTrackingService.cs
@@ -1,6 +1,9 @@
+using Dfe.PlanTech.Core.DataTransferObjects.Sql;
+
 namespace Dfe.PlanTech.Application.Services.Interfaces;
 
 public interface IUserActionTrackingService
 {
     Task RecordActionAsync();
+    Task<SqlUserActionDto?> GetAsync(Guid id);
 }
