--- src/Dfe.PlanTech.Web/Context/Interfaces/IUserActionTrackingService.cs
@@ -1,10 +1,0 @@
-using Dfe.PlanTech.Core.DataTransferObjects.Sql;
-
-namespace Dfe.PlanTech.Web.Context.Interfaces;
-
-public interface IUserActionTrackingService
-{
-    Task RecordAsync();
-    Task<SqlUserActionDto?> GetAsync(Guid id);
-
-}
